chore: replace Main with Main2
Main2 uses more simplified code.
